+++ tests/functional/single_layer_tests/fake_quantize_with_axis.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2023-2025 Intel Corporation
+// Copyright (C) 2023-2026 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
 
@@ -31,7 +31,7 @@ public:
                 std::accumulate(inputStaticShape.begin(), inputStaticShape.end(), 1, std::multiplies<size_t>());
         auto inputTensor = ov::Tensor{ov::element::f32, inputStaticShape};
         auto inputData = inputTensor.data<ov::element_type_traits<ov::element::f32>::value_type>();
-        for (size_t i = 0; i < totalSize; i++) {
+        for (int64_t i = 0; i < totalSize; i++) {
             inputData[i] = (i % 255) - 127;
         }
         inputs = {
